Bytter til Ny OAuthToken-impl, som bruker mer like tokens med refresh for nesten alle(langtlevende)
